+++ CHANGELOG.md
@@ -8,12 +8,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
-## Unreleased
+## 1.14.1 - 2021-04-29
 
 ### Fixed
 
 - Issue when calling `sdk.alerts.update_state()` without specifying a `note` parameter
     would set the existing alert's note's message to the empty string.
+
+### Added
+
+- Custom exception `Py42OrgNotFoundError`.
+
+### Changed
+
+- `sdk.users.get_all()` now raises `Py42OrgNotFoundError` when the given `org_uid`
+    was not found.
+
+- `sdk.users.get_page()` now raises `Py42OrgNotFoundError` when the given `org_uid`
+    was not found.
 
 ## 1.14.0 - 2021-04-20